pkg: Move metrics from sdk pkg
Since the leftover logic in the pkg/sdk was just metrics related.
